Adding WebAIM contrast checker to tools section
I added the WebAIM contrast checker. I think this is the defacto standard (@ericwbailey ?) and we should define our standards of measurement for color contrast.

I also added a note on tools that use color pickers as I noticed huge divergence when using the "Contrast" menubar app to WebAIM or the figma plugin.