include Typography utility class names in documentation for easier reference (#796)
## Context

Whenever I look up typography utility classes, I end up having to inspect the markup in the embedded Storybook to get the actual class name. It would be much easier if the class names were written out in the documentation instead, so they could be copy/pasted directly from the page, without having to take the extra step of inspecting the page source.

## Summary

This PR adds the class names to the documentation. I'm opening a separate PR to add them within the storybooks as well.